chore: use `rolldown` for generating `.d.ts` ⚡️
