+++ packages/@repo/package.config/src/package.config.ts
@@ -34,4 +34,5 @@ export default defineConfig({
     noImplicitSideEffects: 'error',
     noPublishConfigExports: 'error',
   },
+  dts: 'rolldown',
 })
